mc-47 morphic-specific profile: Added conditional logic and test cases for morphic-specifc profile

+++ src/main/java/org/humancellatlas/ingest/study/web/StudyController.java
@@ -9,12 +9,15 @@ import org.humancellatlas.ingest.study.StudyRepository;
 import org.humancellatlas.ingest.study.StudyService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.core.env.Environment;
 import org.springframework.data.rest.webmvc.PersistentEntityResourceAssembler;
 import org.springframework.data.rest.webmvc.RepositoryRestController;
 import org.springframework.hateoas.ExposesResourceFor;
 import org.springframework.hateoas.Resource;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Arrays;
 
 @RepositoryRestController
 @ExposesResourceFor(Study.class)
@@ -27,26 +30,40 @@ public class StudyController {
 
     private final @NonNull StudyRepository studyRepository;
 
+    private final Environment environment;
+
 
     @PostMapping("/studies")
     public ResponseEntity<Resource<?>> registerStudy(@RequestBody final Study study,
                                                      final PersistentEntityResourceAssembler assembler) {
-        Study result = studyService.register(study);
-        return ResponseEntity.ok().body(assembler.toFullResource(result));
+        if (Arrays.asList(environment.getActiveProfiles()).contains("morphic")) {
+            Study result = studyService.register(study);
+            return ResponseEntity.ok().body(assembler.toFullResource(result));
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PatchMapping("/studies/{studyId}")
     public ResponseEntity<Resource<?>> updateStudy(@PathVariable String studyId,
                                                    @RequestBody final ObjectNode patch,
                                                    final PersistentEntityResourceAssembler assembler) {
-        Study updatedStudy = studyService.update(studyId, patch);
-        return ResponseEntity.ok().body(assembler.toFullResource(updatedStudy));
+        if (Arrays.asList(environment.getActiveProfiles()).contains("morphic")) {
+            Study updatedStudy = studyService.update(studyId, patch);
+            return ResponseEntity.ok().body(assembler.toFullResource(updatedStudy));
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @DeleteMapping("/studies/{studyId}")
     public ResponseEntity<Void> deleteStudy(@PathVariable String studyId) {
-        studyService.delete(studyId);
-        return ResponseEntity.noContent().build();
+        if (Arrays.asList(environment.getActiveProfiles()).contains("morphic")) {
+            studyService.delete(studyId);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
 }
